Add argv types to commands
